+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Added
+- [Feature] Create generic `OverlayTrigger` component that intelligently positions an overlay on a page relative to it's children. (#572)
+- [Feature] Create simple `Popover` component. (#462)
 
 ## [17.2.0][17.2.0] - 2016-09-16
 ### Added